requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.7
decorator==5.1.1
Django==4.2.4
django-autoslug==1.9.9
django-crispy-forms==2.0
django-extensions==3.2.3
django-tinymce==3.6.1
easy-thumbnails==2.8.5
executing==1.2.0
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
python-slugify==8.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
